requirements.txt
beautifulsoup4==4.12.2
feedparser==6.0.10
nltk==3.8.1
requests==2.31.0
transformers==4.25.1
torch==2.1.1
python-dateutil==2.8.2
numpy<2.0.0
tqdm==4.64.1
sentencepiece